Big push: User roles can now get edited & the users can be deleted. Filtered out sudo users. Dashboard is filtered out by what is allowed for the user to navigate. Still need to implement more verification.

+++ app/Http/Controllers/HomeController.php
@@ -35,10 +35,11 @@ class HomeController extends Controller
 
     public function destroy($id)
     {
-        dd($id);
-        $user = User::find($id);
-        $user->delete();
-        return redirect('pages.roles');
+        if(isset($id)){
+            $user = User::find($id);
+            $user->delete();
+        }
+        return redirect('/home');
     }
 
     public function update(Request $request)
